Add structure to example so all apps can be explored
