+++ CommandDotNet.Example/Issues/Issue55.cs
@@ -1,0 +1,53 @@
+using System;
+using CommandDotNet.Attributes;
+
+namespace CommandDotNet.Example.Issues
+{
+    public class Issue55 : IDisposable
+    {
+        private readonly string _baseId;
+        private readonly string _host;
+
+        [SubCommand]
+        public Issue55SubCommand Number { get; set; }
+        
+        public Issue55([Option(Inherited = true)]string baseId, [Option(Inherited = true)]string host = "localhost:8080")
+        {
+            this._baseId = baseId;
+            this._host = host;
+            
+            //some more initial work here
+            Console.WriteLine("Initial work...");
+        }
+
+        public void getFirmwareVersion()
+        {
+            Console.WriteLine(this._baseId);
+            Console.WriteLine(this._host);
+        }
+
+        public void Dispose()
+        {
+            Console.WriteLine("Cleanup work ...");
+        }
+    }
+    
+    [ApplicationMetadata(Name = "number")]
+    public class Issue55SubCommand : IDisposable
+    {
+        public Issue55SubCommand()
+        {
+            Console.WriteLine("subcommand constructor executed");
+        }
+        
+        public void Print(int number)
+        {
+            Console.WriteLine(number);
+        }
+        
+        public void Dispose()
+        {
+            Console.WriteLine("Cleanup work from subcommand...");
+        }
+    }
+}
